+++ CHANGELOG.md
@@ -12,6 +12,7 @@ adheres to [Semantic Versioning][semver].
 - A `Snek.Board` module (and submodules) for representing board positions.
 - A `Snek.Ruleset` behaviour module for implementing variations of game rules.
 - A `Snek.Ruleset.Standard` module implementing the Battlesnake Standard rules.
+- A `Snek.Ruleset.Solo` module implementing the Battlesnake Solo rules.
 
 [Unreleased]: https://github.com/xtagon/snek/tree/edge